Refactor webserver into separate file
